pyproject.toml
[project]
name = "aider-v2"
description = "Aider v2 - AI pair programming with improved architecture"
readme = "README.md"
version = "2.0.0"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0.0",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",
    "pydantic>=2.0.0",
    "gitpython>=3.1.0",
    "openai>=1.0.0",
    "anthropic>=0.25.0",
    "litellm>=1.0.0",
    "python-dotenv>=1.0.0",
    "pathspec>=0.11.0",
    "tree-sitter>=0.20.0",
    "pygments>=2.15.0",
    "aiohttp>=3.8.0",
    "pyyaml>=6.0",
    "toml>=0.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
aider-v2 = "aider_v2.cli.main:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["aider_v2*"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true